% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conclique.R
\docType{package}
\name{conclique}
\alias{conclique}
\alias{conclique-package}
\title{conclique: Gibbs sampling for spatial data and concliques}
\description{
Gibbs sampling for spatial data using Markov Random Fields and concliques. 
Currently contains functionality for creating a minimal conclique cover as well as assigning the concliques
to a regular lattice for plotting and simulation purposes. This package is under development and will 
include simulation of spatial data using a conclique-based Gibbs sampler.
}
