% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_single_param.R
\name{gaussian_single_param}
\alias{gaussian_single_param}
\alias{gaussian_single_param_cdf}
\alias{gaussian_single_param_sampler}
\title{Functions for using conclique based GOF test for Gaussian MRF with a single dependence parameter.}
\usage{
gaussian_single_param_sampler(data, params)

gaussian_single_param_cdf(data, params)
}
\arguments{
\item{data}{A list containing two elements, 
sums and nums which contain the sum of the data in each neighborhood as well as the number of locations 
in the neighborhood for each point in the conclique.}

\item{params}{A list of parameter values, rho, kappa, and eta, that parameterize the Gaussian MRF.}
}
\description{
Functions for using conclique based GOF test for Gaussian MRF with a single dependence parameter.
}

