% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neighbors.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Get the neighbors of each location in the lattice.}
\usage{
get_neighbors(lattice, directional = FALSE, grid = NULL)
}
\arguments{
\item{lattice}{The simplified igraph object storing the lattice and dependency structure, ordered by 
location.}

\item{directional}{Indication of if neighborhood needs to be split into North/South, East/West directions. Defaults to FALSE.}

\item{grid}{A grid storing the locations of each point in the lattice. Only necessary is direction = TRUE.}
}
\value{
A data frame N*N by (max # neighbors) + 1, where the first column is the location id of each location in the lattice.
}
\description{
Get the neighbors of each location in the lattice.
}

