% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_statistics.R
\name{gof_statistics}
\alias{gof_statistics}
\title{Get goodness of fit statistics for the spatial residuals of a MRF model.}
\usage{
gof_statistics(residuals, conclique_cover, statistic = c("ks", "cvm"),
  aggregate = c("mean", "max"))
}
\arguments{
\item{residuals}{A vector containing residuals for each location in the lattice, 
output from spatial_residuals.}

\item{conclique_cover}{A list of class "conclique_cover" encoding the locations in each conclique for 
the conclique cover}

\item{statistic}{Which goodness of fit statistic to use, Kolmogorov-Smirnov ("ks") or 
Cramer von Mises ("cvm"). Kolmogorov-Smirnov is the default choice.}

\item{aggregate}{How to aggregate GOF statistics across concliques, mean ("mean") or 
max ("max"). Mean is the default. Can also be a name of a user defined aggregation function.}
}
\description{
Get goodness of fit statistics for the spatial residuals of a MRF model.
}

