% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_4nn_torus.R
\name{lattice_4nn_torus}
\alias{lattice_4nn_torus}
\title{Create a regular lattice of specified size with four-nearest neighborhood dependence structure on a torus}
\usage{
lattice_4nn_torus(dimvec)
}
\arguments{
\item{dimvec}{A 2-dimensional vector specifying the size of the lattice}
}
\description{
Create a regular lattice of specified size with four-nearest neighborhood dependence structure on a torus
}
\examples{
\dontrun{
    lattice <- lattice_4nn_torus(c(6,6))
}
}

