% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_conclique_cover.R
\name{min_conclique_cover}
\alias{min_conclique_cover}
\title{Create a minimal conclique cover for a regular lattice}
\usage{
min_conclique_cover(lattice, max_concliques = list())
}
\arguments{
\item{lattice}{The simplified igraph object storing the lattice and dependency structure, ordered by location}

\item{max_concliques}{The current list of the largest concliques, defaults to empty initial value}
}
\description{
Create a minimal conclique cover for a regular lattice
}
\examples{
\dontrun{
    lattice <- igraph::make_lattice(c(6,6))
    concliques <- min_conclique_cover(lattice)
}
}

