% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_sequential_gibbs}
\alias{run_sequential_gibbs}
\title{Run a sequential Gibbs sampler to sample spatial data given a lattice and neighborhood structure.}
\usage{
run_sequential_gibbs(
  neighbors,
  inits,
  conditional_sampler,
  params,
  n_iter = 100L
)
}
\arguments{
\item{neighbors}{A matrix N*N by (max // neighbors) + 1, where the first column is the location id of each location in the lattice. This could be the result from get_neighbors().
If NULL, will be calculated within the function.}

\item{inits}{Initial values for the lattice, formatted as a grid.}

\item{conditional_sampler}{The string name of a function that has two inputs: 
\itemize{
  \item{data, and}
  \item{params.}
}
There are three built in samplers:
 \itemize{
   \item{"gaussian_single_param" - a Gaussian sampler with a single dependence parameter,}
   \item{"binary_single_param" - a binary sampler with a single dependence parameter, and}
   \item{"binary_two_param" - a binary sampler with two dependence parameters.}
   \item{"binary_two_param_reg" - a binary sampler with two dependence parameters and regression on the horizontal coordinate.}
 }
If the user chooses to write their own sampler in R, they must pass the name of the sampler that is available in the gloabl environment as this parameter.
The input "data" is a list containing two elements, 
sums and nums which contain the sum of the data in each neighborhood as well as the number of locations 
in the neighborhood for each point in the conclique. The input "params" is a list of parameter values. This function returns 
a value sampled from the specified conditional distribution given the data and parameters passed.}

\item{params}{A list of parameters to be passed to the conditional_sampler function}

\item{n_iter}{Number of times to run the Gibbs sampler}
}
\description{
Run a sequential Gibbs sampler to sample spatial data given a lattice and neighborhood structure.
}
