% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampler}
\alias{binary_single_param_sampler}
\alias{binary_two_param_sampler}
\alias{gaussian_single_param_sampler}
\alias{sampler}
\title{Functions for sampling from conditional distributions.}
\usage{
gaussian_single_param_sampler(data, params)

binary_single_param_sampler(data, params)

binary_two_param_sampler(data, params)
}
\arguments{
\item{data}{A list containing two elements, 
sums and nums which contain the sum of the data in each neighborhood as well as the number of locations 
in the neighborhood for each point in the conclique.}

\item{params}{A list of parameter values, rho, kappa, and eta, that parameterize the Gaussian MRF.}
}
\description{
Functions for sampling from conditional distributions.
}

