% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_residuals.R
\name{spatial_residuals}
\alias{spatial_residuals}
\title{Run a conclique-based Gibbs sampler to sample spatial data given a lattice and neighborhood structure.}
\usage{
spatial_residuals(data_grid, lattice, conditional_dsn, params)
}
\arguments{
\item{data_grid}{The grid of data values for each location in the lattice}

\item{lattice}{The simplified igraph object storing the lattice and dependency structure, ordered by 
location}

\item{conditional_dsn}{A function that has two inputs: 
\itemize{
  \item{data, and}
  \item{params.}
}
The data is a data frame containing four columns, vertex, data_value, sum_neighbor, and num_neighbor which encode the values for each location and the neighboring locations of each location 
and the number of neighbors each location has. params is a list of parameter values. This function returns the inverse cdf 
at a value between 0 and 1 from the conditional distribution}

\item{params}{A list of parameters to be passed to the conditional_density function}
}
\description{
Run a conclique-based Gibbs sampler to sample spatial data given a lattice and neighborhood structure.
}

