% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_conclique_gibbs}
\alias{run_conclique_gibbs}
\title{Run a conclique-based Gibbs sampler with a single dependency parameter to sample spatial data given a lattice and neighborhood structure.}
\usage{
run_conclique_gibbs(conclique_cover, neighbors, inits, conditional_sampler,
  params, n_iter = 100L)
}
\arguments{
\item{conclique_cover}{A list of class "conclique_cover" encoding the locations in each conclique for 
the conclique cover}

\item{neighbors}{A list of matrices encoding the neighbors for each point, where the first column of each matrix is the location id of each location in the lattice. This could be the result from get_neighbors().
Include multiple matrices within the list for more complicated neighborhood structures in your corresponding sampler, for example a (two) dependency parameter for N/S, E/W.}

\item{inits}{Initial values for the lattice, formatted as a grid.}

\item{conditional_sampler}{The string name of a function that has three inputs: 
\itemize{
  \item{value,}
  \item{data, and}
  \item{params.}
}
There are three built in samplers:
 \itemize{
   \item{"gaussian_single_param" - a Gaussian sampler with a single dependence parameter,}
   \item{"binary_single_param" - a binary sampler with a single dependence parameter, and}
   \item{"binary_two_param" - a binary sampler with two dependence parameters, and}
 }
If the user chooses to write their own sampler in R, they must pass the name of the sampler that is available in the gloabl environment as this parameter.
Value is a vector of values between 0 and 1. The data is a list containing two element, 
sums and nums which contain the sum of the data in each neighborhood as well as the number of locations 
in the neighborhood for each point in the conclique. params is a list of parameter values. This function returns 
a value sampled from the specified conditional distribution given the data and parameters passed.}

\item{params}{A list of parameters to be passed to the conditional_density function}

\item{n_iter}{Number of times to run the Gibbs sampler}
}
\description{
Run a conclique-based Gibbs sampler with a single dependency parameter to sample spatial data given a lattice and neighborhood structure.
}

