% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_residuals.R
\name{spatial_residuals}
\alias{spatial_residuals}
\title{Get spatial residuals from data given a neighborhood structure and MRF model.}
\usage{
spatial_residuals(data, neighbors, conditional_dsn, params)
}
\arguments{
\item{data}{A vector containing data values for each location in the lattice.}

\item{neighbors}{A matrix N*N by (max # neighbors) + 1, where the first column is the location id of each location in the lattice. This could be the result from get_neighbors().}

\item{conditional_dsn}{A function that has two inputs: 
\itemize{
  \item{data, and}
  \item{params.}
}
The data is a list containing data, sums, and nums, which contains the sampled values for each location, the neighboring locations of each location,
and the number of neighbors each location has, respectively. params is a list of parameter values. This function returns the inverse cdf 
at a value between 0 and 1 from the conditional distribution}

\item{params}{A list of parameters to be passed to the conditional_density function}
}
\description{
Get spatial residuals from data given a neighborhood structure and MRF model.
}

